+++ packages/scaffold/src/partials/w3m-account-default-widget/styles.ts
@@ -42,14 +42,14 @@ export default css`
     height: 48px;
     padding: var(--wui-spacing-xs);
     padding-right: var(--wui-spacing-s);
-    box-shadow: inset 0 0 0 1px var(--wui-gray-glass-002);
-    background-color: var(--wui-gray-glass-002);
+    box-shadow: inset 0 0 0 1px var(--wui-color-gray-glass-002);
+    background-color: var(--wui-color-gray-glass-002);
     border-radius: 24px;
     transaction: background-color 0.2s linear;
   }
 
   .account-button:hover {
-    background-color: var(--wui-gray-glass-005);
+    background-color: var(--wui-color-gray-glass-005);
   }
 
   .avatar-container {
@@ -59,7 +59,7 @@ export default css`
   wui-avatar.avatar {
     width: 32px;
     height: 32px;
-    box-shadow: 0 0 0 2px var(--wui-gray-glass-005);
+    box-shadow: 0 0 0 2px var(--wui-color-gray-glass-005);
   }
 
   wui-avatar.network-avatar {
@@ -69,7 +69,7 @@ export default css`
     left: 100%;
     top: 100%;
     transform: translate(-75%, -75%);
-    box-shadow: 0 0 0 2px var(--wui-gray-glass-005);
+    box-shadow: 0 0 0 2px var(--wui-color-gray-glass-005);
   }
 
   .account-links {
@@ -107,6 +107,6 @@ export default css`
   }
 
   .account-links wui-flex wui-icon svg path {
-    stroke: #47a1ff;
+    stroke: #667dff;
   }
 `