+++ packages/wagmi/src/utils/defaultWagmiReactConfig.ts
@@ -2,10 +2,10 @@ import '@web3modal/polyfills'
 
 import type { CreateConfigParameters, CreateConnectorFn, Config } from 'wagmi'
 import { createConfig } from 'wagmi'
-import { coinbaseWallet, walletConnect } from 'wagmi/connectors'
-
-import { emailConnector } from '../connectors/EmailConnector.js'
+import { coinbaseWallet, walletConnect, injected } from 'wagmi/connectors'
+import { authConnector } from '../connectors/AuthConnector.js'
 import { getTransport } from './helpers.js'
+import type { SocialProvider } from '@web3modal/scaffold-utils'
 
 export type ConfigOptions = Partial<CreateConfigParameters> & {
   chains: CreateConfigParameters['chains']
@@ -13,10 +13,10 @@ export type ConfigOptions = Partial<CreateConfigParameters> & {
   enableEIP6963?: boolean
   enableCoinbase?: boolean
   enableEmail?: boolean
-  /**
-   * Use enableEIP6963 to show all injected wallets
-   * @deprecated
-   */
+  auth?: {
+    socials?: SocialProvider[]
+    showWallets?: boolean
+  }
   enableInjected?: boolean
   enableWalletConnect?: boolean
   metadata: {
@@ -33,8 +33,11 @@ export function defaultWagmiConfig({
   metadata,
   enableCoinbase,
   enableEmail,
+  enableInjected,
+  auth = {
+    showWallets: true
+  },
   enableWalletConnect,
-
   enableEIP6963,
   ...wagmiConfig
 }: ConfigOptions): Config {
@@ -50,6 +53,10 @@ export function defaultWagmiConfig({
     connectors.push(walletConnect({ projectId, metadata, showQrModal: false }))
   }
 
+  if (enableInjected !== false) {
+    connectors.push(injected({ shimDisconnect: true }))
+  }
+
   if (enableCoinbase !== false) {
     connectors.push(
       coinbaseWallet({
@@ -61,8 +68,16 @@ export function defaultWagmiConfig({
   }
 
   // Dissabled by default
-  if (enableEmail === true) {
-    connectors.push(emailConnector({ chains: [...chains], options: { projectId } }))
+  if (enableEmail || auth?.socials) {
+    connectors.push(
+      authConnector({
+        chains: [...chains],
+        options: { projectId },
+        socials: auth?.socials,
+        email: enableEmail,
+        showWallets: auth.showWallets
+      })
+    )
   }
 
   return createConfig({